chore: organised ingestor modules ready for more data source scripts
