+++ services/ingestor/usx/paragraph.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from usxingestor.chapter import Chapter
+    from ingestor.usx.chapter import Chapter
     from manager.logmanager import LogManager
 
 class Paragraph:
