+++ services/ingestor/usx/strongsingestor.py
