+++ services/ingestor/usx/translationnote.py
@@ -3,10 +3,10 @@ import re
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from usxingestor.chapter import Chapter
+    from ingestor.usx.chapter import Chapter
     from manager.logmanager import LogManager
 
-from usxingestor.verse import Verse
+from ingestor.usx.verse import Verse
 
 #region Cases To Handle (both from source and for destination for both cross references and footnotes)
 # 2KI 6:31-7:20
